RUBY-2537 split collection spec for readability
